+++ public/ide/modules/ide/js/IDEController.js
@@ -1,7 +1,7 @@
 /**
  * Created by shenaolin on 16/2/26.
  */
-var ide=angular.module('ide',['ui.bootstrap.contextMenu','colorpicker.module','ui.sortable','btford.modal','ui.bootstrap','ngAnimate','GlobalModule','ui.tree','IDEServices']);
+var ide=angular.module('ide',['ui.bootstrap.contextMenu','colorpicker.module','btford.modal','ui.bootstrap','ngAnimate','GlobalModule','ui.tree','IDEServices']);
 
 
 ide.config(['$compileProvider',
@@ -777,6 +777,7 @@ ide.controller('IDECtrl', [ '$scope','$timeout','$http','$interval', 'ProjectSer
             if(template[key] instanceof Array){
                 //resourcelist
                 template[key].forEach(function(item){
+
                     tempSrc = item.src&&item.src.split('/');
                     tempSrc = tempSrc[tempSrc.length-1];
                     tempSrc = resourceUrl + tempSrc;
@@ -788,10 +789,12 @@ ide.controller('IDECtrl', [ '$scope','$timeout','$http','$interval', 'ProjectSer
                     template[key].texList.forEach(function(tex){
                         if(tex.slices){
                             tex.slices.forEach(function(slice){
-                                tempSrc = slice.imgSrc&&slice.imgSrc.split('/');
-                                tempSrc = tempSrc[tempSrc.length-1];
-                                tempSrc = resourceUrl+tempSrc;
-                                slice.imgSrc = tempSrc;
+                                if(slice.imgSrc!==''){
+                                    tempSrc = slice.imgSrc&&slice.imgSrc.split('/');
+                                    tempSrc = tempSrc[tempSrc.length-1];
+                                    tempSrc = resourceUrl+tempSrc;
+                                    slice.imgSrc = tempSrc;
+                                }
                             })
                         }
                     })