+++ changelog.md
@@ -124,3 +124,15 @@ add:
     5. limit duration time to 5000
 
 
+2017.5.15
+fix:
+    1. copy page bug
+    2. mulitiColor progress's cursor not display in simulator
+    3. remove Slide and Switch widget's action
+    4. change some anmiation logic
+    5. num has decimal in simulator bug
+    6. remove textarea widget's scale ctrl
+    7. widget in default template imgSrc bug
+add:
+    1. add change subLayer zIndex fun
+    2. now can delete multi-files at once